Added a mavenized version of the Drexel University Dragon Toolkit.
The original download is from http://dragon.ischool.drexel.edu. There,
the dragontools sources (this project) as well as all third-party
dependency JAR files can be downloaded. Unfortunately, most dependencies
are not available from public Maven repositories and not even the
sources can be found. For JCoRe, the dragontools are just a dependency
for the BANNER gene tagger. As such, the dependencies required for the
parts that BANNER uses have been added which is primarily the Hepple PoS
tagger. The tagger could be found as a part of GATE core. Some other
dependencies were available as well, but for all cases where no
dependency could be found, the respective classes were just completed
outcommented. Refer to the README.md for more details.