Added capabilities to lookup classpath resources.
We would like to use BANNER with this project and read all resources
from within a JAR file on the classpath. Thus, the FileUtil was modified
to support classpath resources. To the HeppleTagger, URL objects are
passed which by default do not resolve to classpath resources. Thus, a
new protocol Handler "de.julielab.protocols.classpath.Handler" was added
for the "classpath" protocol. The HeppleTagger resource URLs are now
resolved against such classpath resources if they are not accessible as
regular files.