Bug fixes.
Issue: End elements would not always get all the storage keys of their respective start elements because the end elements are not maintained in a map of their own.
Solution: Don't try to keep track of the storage keys for end elements during XMI element extraction but just use a stack to remember the last start element storage key and give it to the elements within the element. We need a stack because some elements may have child elements which in turn have their own storage keys. With the stack we can be sure that each end element really refers to its associated start element.

The code must still be cleaned, there is debug output in there.