Did some refactoring.
Actually, there is some error in production where an XMI ID collision
happens but elementsToWrite.get(newXmiId) is null. I don't quite know
why that is. Tried to refactor the code in a simple way to fix some
ideas I have as to how the error happens. However, this original code
has grown for quite some time and is quite complicated. Perhaps
a rewrite that would be nearer to the formalism described in the
JeDIS paper would be much easier than to try and fix this
spaghetti code.