Created an algorithm to determine elements for omission.
When data is loaded that references other data which hasn't been loaded,
the respective XMI:ID links have to be removed. This can cause
chains of removals, especially in lists. The current algorithm
handles lists in which some nodes are removed and others aren't
by circumventing removed nodes. The XMI:ID links of nodes referencing
them are changed to reference the next non-removed node instead.