Fixed the issue where the XML-entity-resolved string values were put into the encoded data.
With VTD XML it was easy to directly access the underlying byte[] data.
However, VTD XML cannot handle very long `sofaString` attributes
as they occur for PMC full texts.
With StAX, it is virtually impossible to get the attribute value
offsets in the raw byte[] data. For this reason, the new
`AttributeParser` has been written for this purpose.
All tests running.