Add option to combine array-values XPath results for XML ForEach CLI tool.
Version bump of julie-xml-tools to 0.6.3. There is a new possible value that can be given to the Java system property `de.julielab.xmltools.delim` when using the `JulieXMLToolsCLIRecords` tool named `explode`.
Then, array-values XPath results are not just concatenated into a single output cell of the TSV table but each value of the array is combined with all the other column values.