Add new field attribute "completeXml".
If set to "true", the complete XML file contents are the value of that field. We need this for PMC parsing where we use the doc type declaration to find the DTD and search some information in different XPath locations depending on the DTD in the JCoRe PubMed Central parser.